requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]
faiss-cpu
numpy
sentence-transformers
